+++ src/or/hs_service.c
@@ -1875,8 +1875,9 @@ can_service_launch_intro_circuit(hs_service_t *service, time_t now)
                safe_str_client(service->onion_address),
                get_max_intro_circ_per_period(service),
                INTRO_CIRC_RETRY_PERIOD,
-               service->state.num_intro_circ_launched, elapsed_time,
-               INTRO_CIRC_RETRY_PERIOD - elapsed_time);
+               service->state.num_intro_circ_launched,
+               (long int) elapsed_time,
+               (long int) INTRO_CIRC_RETRY_PERIOD - elapsed_time);
       tor_free(msg);
     }
   }
